D8P2 Finally done
The holidays doesn't bring a lot of focus to the Byrum household, and it
shows in how long this took. I had a few minor issues in the code that
kept pushing me off the right solution, which I had in the algorithm.
I'd have learned that earlier if I had broken the problem down a bit
more and written a few basic unit tests. The old lessons don't stop
applying just because it's for fun! :)